requirements.txt
feedparser>=6.0
requests>=2.31
anthropic>=0.40
openai>=1.50
jinja2>=3.1
pyyaml>=6.0
python-dotenv>=1.0
pydub>=0.25
